Fix #56, Add RwLock API & its POSIX implementation
Create a shared RwLock API, along with a POSIX implementation and the
associated test cases and stubs. Add configuration option to disable
RwLocks.

Since some persistent data structures within cFS are read-heavy, it
would be wise to take advantage of a readers-writer lock over a mutex
for synchronization for performance reasons.