Fix cFS/cFS#280, add DNS support to OS_SocketAddrFromString
Adds an alternative implementation of OS_SocketAddrFromString that uses
the getaddrinfo() system call to look up the name in DNS (or whatever
network name service is configured).

Without changing the API, this can only return the first address.

The existing (non-DNS) option is retained for RTOSs that may not have
this feature.  The DNS version is only enabled on posix for now.